+++ ports/servoshell/desktop/embedder.rs
@@ -6,8 +6,8 @@
 
 use net::protocols::ProtocolRegistry;
 use servo::compositing::windowing::EmbedderMethods;
-use servo::embedder_traits::{EmbedderProxy, EventLoopWaker};
 use servo::servo_config::pref;
+use servo::{EmbedderProxy, EventLoopWaker};
 use webxr::glwindow::GlWindowDiscovery;
 #[cfg(target_os = "windows")]
 use webxr::openxr::OpenXrDiscovery;
