+++ ports/servoshell/desktop/minibrowser.rs
@@ -21,12 +21,11 @@ use glow::NativeFramebuffer;
 use log::{trace, warn};
 use servo::base::id::WebViewId;
 use servo::compositing::windowing::EmbedderEvent;
-use servo::script_traits::TraversalDirection;
 use servo::servo_geometry::DeviceIndependentPixel;
 use servo::servo_url::ServoUrl;
 use servo::webrender_api::units::DevicePixel;
 use servo::webrender_traits::SurfmanRenderingContext;
-use servo::TopLevelBrowsingContextId;
+use servo::{TopLevelBrowsingContextId, TraversalDirection};
 use winit::event::{ElementState, MouseButton, WindowEvent};
 use winit::event_loop::ActiveEventLoop;
 use winit::window::Window;
